Added tokenized output validations.

+++ tests/test_tokenization.py
@@ -39,9 +39,9 @@ SCHEMAS_SCHEMA = {
 
 SEQ_SCHEMA = {
     "patient_id": NORMALIZED_MEDS_SCHEMA["patient_id"],
-    "code": pl.List(pl.List(pl.Float64)),
+    "code": pl.List(pl.List(pl.UInt8)),
     "numeric_value": pl.List(pl.List(NORMALIZED_MEDS_SCHEMA["numeric_value"])),
-    "time_delta_days": pl.List(pl.Float64),
+    "time_delta_days": pl.List(pl.Float32),
 }
 
 TRAIN_0_TIMES = [
